test fixes and refactoring
