--- MDANSE/Tests/UnitTests/select_atoms.py
@@ -1,3 +1,0 @@
-selection = [
-    at for at in chemicalSystem.atom_list() if at.element.strip() == "phosphorus"
-]
